requirements.txt
flask_ngrok 
flask_bootstrap
transformers
torch
